requirements.txt
opencv-python
numpy
requests
sentence-transformers
torch
transformers
overpy
Pillow
wikipedia
open_clip_torch
